Refactor HttpUtil and OnepayHttpUtil
Move those classes to the new package cl.transbank.util and rename
OnepayHttpUtil to HttpUtilImpl in order to support the new webpay
Rest implementation.
