+++ src/main/java/cl/transbank/onepay/ApiBaseResource.java
@@ -1,6 +1,8 @@
 package cl.transbank.onepay;
 
 import cl.transbank.onepay.util.*;
+import cl.transbank.util.HttpUtil;
+import cl.transbank.util.HttpUtilImpl;
 import lombok.NonNull;
 
 import java.io.IOException;
@@ -10,7 +12,7 @@ public abstract class ApiBaseResource {
     private static RequestBuilder requestBuilder = OnepayRequestBuilder.getInstance();
     private static JsonUtil jsonUtil = OnepayJsonUtil.getInstance();
     private static SignUtil signUtil = OnepaySignUtil.getInstance();
-    private static HttpUtil httpUtil = OnepayHttpUtil.getInstance();
+    private static HttpUtil httpUtil = HttpUtilImpl.getInstance();
 
     protected static RequestBuilder getRequestBuilder() {
         return requestBuilder;
@@ -44,12 +46,12 @@ public abstract class ApiBaseResource {
         ApiBaseResource.httpUtil = httpUtil;
     }
 
-    protected static String request(@NonNull URL url, OnepayHttpUtil.RequestMethod method, @NonNull String query)
+    protected static String request(@NonNull URL url, HttpUtilImpl.RequestMethod method, @NonNull String query)
             throws IOException {
         return getHttpUtil().request(url, method, query);
     }
 
-    public String request(@NonNull URL url, OnepayHttpUtil.RequestMethod method, @NonNull String query,
+    public String request(@NonNull URL url, HttpUtilImpl.RequestMethod method, @NonNull String query,
                           HttpUtil.ContentType contentType) throws IOException {
         return getHttpUtil().request(url, method, query, contentType);
     }
