+++ src/main/java/cl/transbank/util/HttpUtil.java
@@ -1,4 +1,4 @@
-package cl.transbank.onepay.util;
+package cl.transbank.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import java.io.IOException;
 import java.net.URL;
 
 public interface HttpUtil {
-    String request(URL url, OnepayHttpUtil.RequestMethod method, String query) throws IOException;
-    String request(URL url, OnepayHttpUtil.RequestMethod method, String query, HttpUtil.ContentType contentType)
+    String request(URL url, HttpUtilImpl.RequestMethod method, String query) throws IOException;
+    String request(URL url, HttpUtilImpl.RequestMethod method, String query, HttpUtil.ContentType contentType)
             throws IOException;
 
     @AllArgsConstructor
